requirements/base.txt
Django>=3.1,<3.2
wagtail>=2.11,<2.12
python-decouple==3.4
